requirements.txt
streamlit
pandas
numpy
torch
torchvision
transformers
scikit-learn
tensorflow
keras
pyvi
pillow
streamlit-option-menu
